use (s)size_t types for streams

+++ streams.h
@@ -22,11 +22,11 @@
 
 typedef struct {
     char *buffer;
-    int buflen;
-    int current;
+    size_t buflen;
+    size_t current;
 } Stream;
 
-extern Stream *new_stream(int size);
+extern Stream *new_stream(size_t size);
 extern void stream_add_char(Stream *, char);
 extern void stream_delete_char(Stream *);
 extern void stream_add_string(Stream *, const char *);
@@ -34,7 +34,7 @@ extern void stream_printf(Stream *, const char *,...) FORMAT(printf,2,3);
 extern void free_stream(Stream *);
 extern char *stream_contents(Stream *);
 extern char *reset_stream(Stream *);
-extern int stream_length(Stream *);
+extern size_t stream_length(Stream *);
 
 #include "exceptions.h"
